requirements.txt
mysql-connector-python
google-generativeai
streamlit
streamlit_mic_recorder
